+++ donation-service/src/main/java/com/gohelpfund/api/v1/donation_service/wallets/services/promise/PromiseWalletDetailsService.java
@@ -67,7 +67,7 @@ public class PromiseWalletDetailsService {
 
         PromiseWalletDetails promiseWalletDetails = new PromiseWalletDetails();
 
-        int balance = 0;
+        double balance = 0;
         if (type.equalsIgnoreCase("fundraiser")){
             balance = 1000000;
         }
